Rename framework subpackage to torch
