+++ src/pydvl/influence/torch/__init__.py
@@ -1,5 +1,6 @@
 from .torch_differentiable import (
     TorchTwiceDifferentiable,
+    as_tensor,
     solve_arnoldi,
     solve_batch_cg,
     solve_linear,
