+++ src/pydvl/influence/torch/functional.py
