+++ src/pydvl/influence/torch/util.py
